+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * Add `submit_for_settlement` option to `Subscription.retry_charge`
 * Add `options` -> `paypal` -> `description` for creating and updating subscriptions
 * Add Braintree::Dispute.find
+* Add braintree.Dispute.accept
+* Add braintree.Dispute.finalize
+* Add braintree.Dispute.find
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`